refactor menu category data
